Move methods in core/reshape/util.py to where they are used
