+++ pandas/tests/indexes/multi/test_util.py
@@ -6,7 +6,7 @@ from pandas import (
     date_range,
 )
 import pandas._testing as tm
-from pandas.core.reshape.util import cartesian_product
+from pandas.core.indexes.multi import cartesian_product
 
 
 class TestCartesianProduct:
